refactor(agents): reorganize server-agent toolset into an internal tool capability
- Rename the main agent's delegation tool call_server_agent -> call_subagent
- Move the networkless GraphQL toolset into capabilities/tools/internal as
  RunGraphQLQueryToolset wrapped by a new RunGraphQLQueryCapability, alongside
  CallSubAgentCapability (capabilities exported, toolsets kept internal)
- Drop the empty ServerAgentDependencies type; the server agent now uses None deps
- Flatten the graphql package into a single server_agents.py module and build the
  agent via the capability + OpenInferenceCapabilityWrapper
- Extract the server-agent system prompt into an injectable AgentPrompts template
  (server_agent/SERVER_AGENT_INSTRUCTIONS.xml.j2)
- Make DataLoaders/CacheForDataLoaders runtime imports in api/context.py
