--- src/phoenix/server/agents/capabilities/server_agents/__init__.py
@@ -1,9 +1,0 @@
-from phoenix.server.agents.capabilities.server_agents.capability import (
-    ServerAgentCapability,
-)
-from phoenix.server.agents.capabilities.server_agents.toolset import ServerAgentToolset
-
-__all__ = [
-    "ServerAgentCapability",
-    "ServerAgentToolset",
-]
