+++ src/phoenix/server/agents/capabilities/tools/internal/__init__.py
@@ -1,0 +1,11 @@
+from phoenix.server.agents.capabilities.tools.internal.call_subagent import (
+    CallSubAgentCapability,
+)
+from phoenix.server.agents.capabilities.tools.internal.run_graphql_query import (
+    RunGraphQLQueryCapability,
+)
+
+__all__ = [
+    "CallSubAgentCapability",
+    "RunGraphQLQueryCapability",
+]
