--- src/phoenix/server/agents/graphql/__init__.py
@@ -1,7 +1,0 @@
-from phoenix.server.agents.graphql.server_agent import build_server_agent
-from phoenix.server.agents.graphql.types import ServerAgentDependencies
-
-__all__ = [
-    "ServerAgentDependencies",
-    "build_server_agent",
-]
